Add ability to set BGP local_pref on transit
As the Coloclue network expands, we are starting to do remote peering at
SwissIX, FranceIX and DE-CIX. To do this correctly, and to avoid sending
traffic to far away places just to see it come back again, we have
decided to use BGP local preferences to steer egress traffic from
AS8283.

Before this change, localpref was 100 for any sessions, and we solely
relied on as-path length to make the routing decision. Now that we have
a mixture of 10G ports (speedix, lsix, frysix), 100M ports (decix,
franceix, swissix) and transit ports (1G/10G), it becomes more important
to do fine grained traffic engineering.

This change adds the ability to set bgp_local_pref for transit sessions.
By default the localpref stays 100, so this change is non intrusive.

If we add transit.transit1.sessions.session1.bgp_local_pref to the YAML
map, we can then override the localpref on that particular session.

TESTED:

1. I run the transit generator without any changes to vars/transit.yml.
   And I confirm to see no local_pref changes:

```
$ router=eunetworks-2.router.nl.coloclue.net
$ ./gentool -4 -y vars/generic.yml vars/${router}.yml vars/transit.yml -t templates/transit.j2 -o /opt/router-staging/${router}/transit-ipv4.conf
$ grep -c local_pref /opt/router-staging/${router}/transit-ipv4.conf
0
```

2. Then, I added the keyword to one BGP transit session with Fusix
   on eunetworks-2, and confirm that now the default localpref is set:

```
protocol bgp transit_fusix_1 from transit {
    description "Fusix Networks B.V. session 1";
    neighbor 37.139.140.26 as 57866;
    include "ebgp_state.conf";
    import where transit_import(15);
    export where transit_fusix_export(0);
    default bgp_local_pref 70;
}
```