Lower local pref to 0 on received prefixes when graceful shutdown is enabled
We should lower the local pref to 0 on received prefixes when graceful
shutdown is enabled, so we also don't send any traffic towards peers we are
going to shut down the BGP session. We shouldn't only add the graceful shutdown
community on our exports. This to make the transition to different
uplinks even smoother